+++ src/index.mjs
@@ -1,5 +1,0 @@
-import Middleware from './middleware';
-
-export { Middleware };
-
-export default Middleware;