+++ test/benchmark.bench.js
@@ -1,6 +1,6 @@
-const { Middleware } = require('..');
+const { MiddlewareStatus, compose } = require('..');
 
-suite('Chain', () => {
+suite('MiddlewareStatus', () => {
 	set('type', 'adaptive');
 	set('mintime', 1000);
 	set('delay', 100);
@@ -18,10 +18,36 @@ suite('Chain', () => {
 
 		const middlewares = Array(count).fill(fn);
 
-		const middleware = new Middleware(middlewares);
+		const middleware = new MiddlewareStatus(middlewares);
 
 		bench(`(fn * ${count})`, (done) => {
 			middleware.run({}).then(done, done);
+		});
+	}
+});
+
+suite('Compose', () => {
+	set('type', 'adaptive');
+	set('mintime', 1000);
+	set('delay', 100);
+
+	const logic = async () => true;
+
+	const fn = async (ctx, next) => {
+		await logic();
+		await next();
+		await logic();
+	};
+
+	for (let exp = 0; exp <= 10; exp += 1) {
+		const count = 2 ** exp;
+
+		const middlewares = Array(count).fill(fn);
+
+		const middleware = compose(middlewares, () => {});
+
+		bench(`(fn * ${count})`, (done) => {
+			middleware({}).then(done, done);
 		});
 	}
 });